+++ packages/bignumber/lib.esm/_version.d.ts
@@ -1,1 +1,1 @@
-export declare const version = "bignumber/5.0.1";
+export declare const version = "bignumber/5.0.2";