+++ packages/providers/lib.esm/url-json-rpc-provider.d.ts
@@ -1,10 +1,12 @@
 import { Network, Networkish } from "@ethersproject/networks";
 import { ConnectionInfo } from "@ethersproject/web";
 import { JsonRpcProvider, JsonRpcSigner } from "./json-rpc-provider";
-export declare abstract class UrlJsonRpcProvider extends JsonRpcProvider {
+export declare class StaticJsonRpcProvider extends JsonRpcProvider {
+    detectNetwork(): Promise<Network>;
+}
+export declare abstract class UrlJsonRpcProvider extends StaticJsonRpcProvider {
     readonly apiKey: any;
     constructor(network?: Networkish, apiKey?: any);
-    detectNetwork(): Promise<Network>;
     _startPending(): void;
     getSigner(address?: string): JsonRpcSigner;
     listAccounts(): Promise<Array<string>>;