+++ packages/signing-key/lib.esm/_version.d.ts
@@ -1,1 +1,1 @@
-export declare const version = "signing-key/5.0.1";
+export declare const version = "signing-key/5.0.2";